#ILEX-6 Separate seach from header
